Change member function casing to camel case
This chance is in accordance with [Arduino API guidelines](https://www.arduino.cc/en/Reference/APIStyleGuide).

Free functions were also slated to be changed, but cannot be changed currently due to a conflict with some macro definitions within the Arduino corelib itself. This may be resolved later through use of pragmas, by accepting the limitation and sticking to pascal case, or by influencing a change in the Arduino library that eliminates the macro problem altogether.